Remove useMenuTrigger hook
